refactor: remove useSlideGroup composable
