+++ packages/vuetify/src/components/VSlideGroup/VSlideGroup.tsx
@@ -1,24 +1,65 @@
 // Styles
 import './VSlideGroup.sass'
 
+// Components
+import { VFadeTransition } from '@/components/transitions'
+import { VIcon } from '@/components/VIcon'
+
 // Composables
-import { makeGroupProps } from '@/composables/group'
-import { useSlideGroup } from '@/composables/slideGroup'
+import { makeGroupProps, useGroup } from '@/composables/group'
+import { useResizeObserver } from '@/composables/resizeObserver'
+import { useRtl } from '@/composables/rtl'
 import { makeTagProps } from '@/composables/tag'
 
 // Utilities
-import { defineComponent, useRender } from '@/util'
-import { computed } from 'vue'
+import { computed, ref, watch, watchEffect } from 'vue'
+import { clamp, defineComponent, useRender } from '@/util'
+import { bias, calculateCenteredOffset, calculateUpdatedOffset } from './helpers'
+
+// Types
+import type { InjectionKey } from 'vue'
+import type { GroupProvide } from '@/composables/group'
+import { useDisplay } from '@/composables'
+
+export const VSlideGroupSymbol: InjectionKey<GroupProvide> = Symbol.for('vuetify:v-slide-group')
 
 export const VSlideGroup = defineComponent({
   name: 'VSlideGroup',
 
   props: {
-    symbol: null,
+    activeClass: {
+      type: String,
+      default: 'v-slide-item--active',
+    },
+    centerActive: Boolean,
+    direction: {
+      type: String,
+      default: 'horizontal',
+    },
+    symbol: {
+      type: null,
+      default: VSlideGroupSymbol,
+    },
+    nextIcon: {
+      type: String,
+      default: '$next',
+    },
+    prevIcon: {
+      type: String,
+      default: '$prev',
+    },
+    showArrows: {
+      type: [Boolean, String],
+      validator: (v: any) => (
+        typeof v === 'boolean' || [
+          'always',
+          'desktop',
+          'mobile',
+        ].includes(v)
+      ),
+    },
     ...makeTagProps(),
-    ...makeGroupProps({
-      mandatory: true,
-    }),
+    ...makeGroupProps(),
   },
 
   emits: {
@@ -26,59 +67,253 @@ export const VSlideGroup = defineComponent({
   },
 
   setup (props, { slots }) {
-    const {
-      containerRef,
-      contentRef,
-      contentStyles,
-      next,
-      prev,
-      select,
-      selected,
-      isSelected,
-    } = useSlideGroup(props, props.symbol)
+    const { isRtl } = useRtl()
+    const { mobile } = useDisplay()
+    const group = useGroup(props, props.symbol)
+    const isOverflowing = ref(false)
+    const scrollOffset = ref(0)
+    const containerSize = ref(0)
+    const contentSize = ref(0)
+    const isHorizontal = computed(() => props.direction === 'horizontal')
+
+    const { resizeRef: containerRef, contentRect: containerRect } = useResizeObserver()
+    const { resizeRef: contentRef, contentRect } = useResizeObserver()
+
+    watchEffect(() => {
+      const sizeProperty = isHorizontal.value ? 'width' : 'height'
+
+      containerSize.value = containerRect.value?.[sizeProperty] ?? 0
+      contentSize.value = contentRect.value?.[sizeProperty] ?? 0
+
+      isOverflowing.value = containerSize.value + 1 < contentSize.value
+    })
+
+    watch(group.selected, selected => {
+      if (!selected.length || !contentRef.value) return
+
+      const index = group.items.value.findIndex(item => item.id === selected[selected.length - 1])
+
+      // TODO: Is this too naive? Should we store element references in group composable?
+      const selectedElement = contentRef.value.children[index] as HTMLElement
+
+      if (index === 0 || !isOverflowing.value) {
+        scrollOffset.value = 0
+      } else if (props.centerActive) {
+        scrollOffset.value = calculateCenteredOffset({
+          selectedElement,
+          containerSize: containerSize.value,
+          contentSize: contentSize.value,
+          isRtl: isRtl.value,
+          isHorizontal: isHorizontal.value,
+        })
+      } else if (isOverflowing.value) {
+        scrollOffset.value = calculateUpdatedOffset({
+          selectedElement,
+          containerSize: containerSize.value,
+          contentSize: contentSize.value,
+          isRtl: isRtl.value,
+          currentScrollOffset: scrollOffset.value,
+          isHorizontal: isHorizontal.value,
+        })
+      }
+    })
+
+    const disableTransition = ref(false)
+
+    let startTouch = 0
+    let startOffset = 0
+
+    function onTouchstart (e: TouchEvent) {
+      const sizeProperty = isHorizontal.value ? 'clientX' : 'clientY'
+      startOffset = scrollOffset.value
+      startTouch = e.touches[0][sizeProperty]
+      disableTransition.value = true
+    }
+
+    function onTouchmove (e: TouchEvent) {
+      if (!isOverflowing.value) return
+
+      const sizeProperty = isHorizontal.value ? 'clientX' : 'clientY'
+      scrollOffset.value = startOffset + startTouch - e.touches[0][sizeProperty]
+    }
+
+    function onTouchend (e: TouchEvent) {
+      const maxScrollOffset = contentSize.value - containerSize.value
+
+      if (isRtl.value) {
+        if (scrollOffset.value > 0 || !isOverflowing.value) {
+          scrollOffset.value = 0
+        } else if (scrollOffset.value <= -maxScrollOffset) {
+          scrollOffset.value = -maxScrollOffset
+        }
+      } else {
+        if (scrollOffset.value < 0 || !isOverflowing.value) {
+          scrollOffset.value = 0
+        } else if (scrollOffset.value >= maxScrollOffset) {
+          scrollOffset.value = maxScrollOffset
+        }
+      }
+
+      disableTransition.value = false
+    }
+
+    function onScroll () {
+      containerRef.value && (containerRef.value.scrollLeft = 0)
+    }
+
+    function onFocusin (e: FocusEvent) {
+      if (!isOverflowing.value || !contentRef.value) return
+
+      // Focused element is likely to be the root of an item, so a
+      // breadth-first search will probably find it in the first iteration
+      for (const el of e.composedPath()) {
+        for (const item of contentRef.value.children) {
+          if (item === el) {
+            scrollOffset.value = calculateUpdatedOffset({
+              selectedElement: item as HTMLElement,
+              containerSize: containerSize.value,
+              contentSize: contentSize.value,
+              isRtl: isRtl.value,
+              currentScrollOffset: scrollOffset.value,
+              isHorizontal: isHorizontal.value,
+            })
+            return
+          }
+        }
+      }
+    }
+
+    function scrollTo (location: 'prev' | 'next') {
+      const sign = isRtl.value ? -1 : 1
+      const newAbosluteOffset = sign * scrollOffset.value +
+        (location === 'prev' ? -1 : 1) * containerSize.value
+
+      scrollOffset.value = sign * clamp(newAbosluteOffset, 0, contentSize.value - containerSize.value)
+    }
+
+    const contentStyles = computed(() => {
+      const scrollAmount = scrollOffset.value <= 0
+        ? bias(-scrollOffset.value)
+        : scrollOffset.value > contentSize.value - containerSize.value
+          ? -(contentSize.value - containerSize.value) + bias(contentSize.value - containerSize.value - scrollOffset.value)
+          : -scrollOffset.value
+
+      return {
+        transform: `translate${isHorizontal.value ? 'X' : 'Y'}(${scrollAmount}px)`,
+        transition: disableTransition.value ? 'none' : '',
+        willChange: disableTransition.value ? 'transform' : '',
+      }
+    })
 
     const slotProps = computed(() => ({
-      next,
-      prev,
-      select,
-      isSelected,
+      next: group.next,
+      prev: group.prev,
+      select: group.select,
+      isSelected: group.isSelected,
     }))
+
+    const hasAffixes = computed(() => {
+      switch (props.showArrows) {
+        // Always show arrows on desktop & mobile
+        case 'always': return true
+
+        // Always show arrows on desktop
+        case 'desktop': return mobile.value
+
+        // Show arrows on mobile when overflowing.
+        // This matches the default 2.2 behavior
+        case true: return isOverflowing.value || Math.abs(scrollOffset.value) > 0
+
+        // Always show on mobile
+        case 'mobile': return (
+          mobile.value ||
+          (isOverflowing.value || Math.abs(scrollOffset.value) > 0)
+        )
+
+        // https://material.io/components/tabs#scrollable-tabs
+        // Always show arrows when
+        // overflowed on desktop
+        default: return (
+          mobile.value &&
+          (isOverflowing.value || Math.abs(scrollOffset.value) > 0)
+        )
+      }
+    })
+
+    const hasPrev = computed(() => {
+      return hasAffixes.value && scrollOffset.value > 0
+    })
+
+    const hasNext = computed(() => {
+      if (!hasAffixes.value) return false
+
+      // Check one scroll ahead to know the width of right-most item
+      return contentSize.value > Math.abs(scrollOffset.value) + containerSize.value
+    })
 
     useRender(() => (
       <props.tag
         class={[
           'v-slide-group',
+          {
+            'v-slide-group--vertical': !isHorizontal.value,
+            'v-slide-group--has-affixes': hasAffixes.value,
+            'v-slide-group--is-overflowing': isOverflowing.value,
+          },
         ]}
       >
-        { slots.prepend && (
-          <div class="v-slide-group__prepend">
-            { slots.prepend(slotProps.value) }
-          </div>
-        )}
+        <div
+          class={[
+            'v-slide-group__prev',
+            { 'v-slide-group__prev--disabled': !hasPrev.value },
+          ]}
+          onClick={ () => scrollTo('prev') }
+        >
+          { slots.prev?.(slotProps.value) ?? (
+            <VFadeTransition>
+              <VIcon icon={ props.prevIcon }></VIcon>
+            </VFadeTransition>
+          )}
+        </div>
 
         <div
           ref={ containerRef }
           class="v-slide-group__container"
+          onScroll={ onScroll }
         >
           <div
             ref={ contentRef }
             class="v-slide-group__content"
             style={ contentStyles.value }
+            onTouchstart={ onTouchstart }
+            onTouchmove={ onTouchmove }
+            onTouchend={ onTouchend }
+            onFocusin={ onFocusin }
           >
             { slots.default?.(slotProps.value) }
           </div>
         </div>
 
-        { slots.append && (
-          <div class="v-slide-group__append">
-            { slots.append(slotProps.value) }
-          </div>
-        )}
+        <div
+          class={[
+            'v-slide-group__next',
+            { 'v-slide-group__next--disabled': !hasNext.value },
+          ]}
+          onClick={ () => scrollTo('next') }
+        >
+          { slots.next?.(slotProps.value) ?? (
+            <VFadeTransition>
+              <VIcon icon={ props.nextIcon }></VIcon>
+            </VFadeTransition>
+          )}
+        </div>
       </props.tag>
     ))
 
     return {
-      selected,
+      selected: group.selected,
+      scrollTo,
+      scrollOffset,
     }
   },
 })
